Refactor

+++ pdfsdk/src/main/java/com/hungknow/pdfsdk/utils/Utils.kt
@@ -5,8 +5,8 @@ import android.util.TypedValue
 
 class Utils private constructor() {
     companion object {
-        fun getDP(displayMetrics: DisplayMetrics, dp: Float): Float {
-            return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, displayMetrics)
+        fun getDP(displayMetrics: DisplayMetrics, dp: Int): Int {
+            return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp.toFloat(), displayMetrics).toInt()
         }
     }
 }
